fix: GoalPrognosisCalculator Typkonflikt (Float? elvis Double-Literal -> uneindeutiger Typ)
weeklyTargetLoss = profile.weeklyTargetLossKg ?: 0.5 ergab wegen Float?/Double-Mix keinen
klaren numerischen Typ mehr, dadurch schlugen alle Folgeoperationen (-weeklyTargetLoss,
Division, Vergleich) fehl. Fix: explizit auf Double normalisieren.

+++ app/src/main/kotlin/ch/nutrisnap/app/domain/GoalPrognosisCalculator.kt
@@ -57,7 +57,7 @@ class GoalPrognosisCalculator(
     suspend fun calculatePrognosis(): GoalPrognosis? {
         val profile = userProfileRepository.get().first()
         val targetWeight = profile.targetWeightKg ?: return null
-        val weeklyTargetLoss = profile.weeklyTargetLossKg ?: 0.5
+        val weeklyTargetLoss = (profile.weeklyTargetLossKg ?: 0.5f).toDouble()
         val currentWeight = weightRepository.getLatest()?.weightKg ?: return null
 
         val weightToLose = currentWeight - targetWeight
